requirements.txt
cerberus
pyyaml

# google-api-python-client
# google-auth-httplib2
# google-auth-oauthlib

pytest
pytest-instafail
flake8
mypy
black
isort